requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
httpx==0.27.0

# Azure OpenAI
openai==1.12.0

# Image Processing
Pillow==10.2.0

# Environment Variables
python-dotenv==1.0.0

# CORS support
starlette==0.35.1

# Optional: For production
gunicorn==21.2.0
